THE WALL
I'm not Donald Trump but i did make a wall and if the Mexicans (the player) touches THE WALL, they will die.

PS.  the wall will only follow the player so don't forget to assign in WallParent at chasing wall script who the player is.

PPS. also the planforms will disappear! keep in mind, you have to add the tag to the platform